+++ frontend/src/components/Common/ProgressBar.tsx
