+++ frontend/src/components/Test/TestTemplate.tsx
@@ -3,10 +3,10 @@ import styled, { css } from 'styled-components';
 import ChatTemplate from './Chat/ChatTemplate';
 import ResultTemplate from './Result/ResultTemplate';
 import StatTemplate from './Statistics/StatTemplate';
-import Modal from '../_Basic/Modal';
-import ScrollModal from '../_Basic/ScrollModal';
-import SideBar from '../_Basic/SideBar';
-import ProgressBar from '../_Basic/ProgressBar';
+import Modal from '../Common/Modal';
+import ScrollModal from '../Common/ScrollModal';
+import SideBar from '../Common/SideBar';
+import ProgressBar from '../Common/ProgressBar';
 import Backdrop from '@material-ui/core/Backdrop';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import { makeStyles } from '@material-ui/core/styles';
